update when sec equal to 0
