+++ frontend/src/App.js
@@ -2,25 +2,31 @@ import logo from "./logo.svg";
 import "./App.css";
 import ConsentPage from "./components/ConsentPage";
 import Timer from "./components/Timer";
+import { useState, useEffect } from "react";
 
 function App() {
+  //const [record, setRecord] = useState([]);
+  const [sec, setSec] = useState(10);
+  const [background, setbackground] = useState("blue");
+  const [stop, setStop] = useState(true);
+
+  useEffect(() => {
+    if (sec <= 5) {
+      setbackground("red");
+    }
+    const timer = setTimeout(() => {
+      setSec(sec - 1);
+    }, 1000);
+
+    if (sec === 0) {
+      clearTimeout(timer);
+      //setSec("You finish!!");
+    }
+  });
+
   return (
     <div className="App">
-      {/* <header className="App-header">
-        <img src={logo} className="App-logo" alt="logo" />
-        <p>
-          Edit <code>src/App.js</code> and save to reload.
-        </p>
-        <a
-          className="App-link"
-          href="https://reactjs.org"
-          target="_blank"
-          rel="noopener noreferrer"
-        >
-          Learn React
-        </a>
-      </header> */}
-      <Timer />
+      <Timer sec={sec} background={background} />
       <ConsentPage />
     </div>
   );
